model training stable

+++ llama/llama/generation_pretokenized.py
@@ -115,7 +115,9 @@ class Llama2:
         )
         tokenizer = Tokenizer(model_path=tokenizer_path)
         model_args.vocab_size = tokenizer.n_words
-        torch.set_default_tensor_type(torch.cuda.HalfTensor)
+        #TODO try things here
+        torch.set_default_tensor_type(torch.cuda.FloatTensor)
+        # torch.set_default_tensor_type(torch.cuda.HalfTensor)
         model = Transformer(model_args)
         model.load_state_dict(checkpoint, strict=False)
         print(f"Loaded in {time.time() - start_time:.2f} seconds")
